+++ diginode-setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#           Name:  DigiNode Setup v0.8.7
+#           Name:  DigiNode Setup v0.8.8
 #
 #        Purpose:  Install and manage a DigiByte Node and DigiAsset Node via the linux command line.
 #          
@@ -10263,24 +10263,24 @@ if [ "$IPFS_DO_INSTALL" = "YES" ]; then
         if [ "$use_ipfs_server_profile" = "yes" ]; then
             sudo -u $USER_ACCOUNT ipfs init -p server
         elif [ "$use_ipfs_server_profile" = "no" ]; then
-            if [ "$IS_RPI" = "YES" ]; then
-                printf "%b Raspberry Pi Detected! Initializing IPFS daemon with the lowpower profile.\\n" "${INFO}"
-                sudo -u $USER_ACCOUNT ipfs init --profile=lowpower
-            else
+#            if [ "$IS_RPI" = "YES" ]; then
+#                printf "%b Raspberry Pi Detected! Initializing IPFS daemon with the lowpower profile.\\n" "${INFO}"
+#                sudo -u $USER_ACCOUNT ipfs init --profile=lowpower
+#            else
                 # Just in case we are are DigiAsset Node Mode ONLY and we never performed the Pi checks
                 # Look for any mention of 'Raspberry Pi' so we at least know it is a Pi 
-                pigen=$(tr -d '\0' < /proc/device-tree/model | grep -Eo "Raspberry Pi" || echo "")
-                if [[ $pigen == "Raspberry Pi" ]]; then
-                    IS_RPI="YES"
-                fi
-                if [ "$IS_RPI" = "YES" ]; then
-                    printf "%b We are in DigiAsset Mode ONLY.\\n" "${INFO}"
-                    printf "%b Raspberry Pi Detected! Initializing IPFS daemon with the lowpower profile.\\n" "${INFO}"
-                    sudo -u $USER_ACCOUNT ipfs init --profile=lowpower
-                else
+#                pigen=$(tr -d '\0' < /proc/device-tree/model | grep -Eo "Raspberry Pi" || echo "")
+#                if [[ $pigen == "Raspberry Pi" ]]; then
+#                    IS_RPI="YES"
+#                fi
+#                if [ "$IS_RPI" = "YES" ]; then
+#                    printf "%b We are in DigiAsset Mode ONLY.\\n" "${INFO}"
+#                    printf "%b Raspberry Pi Detected! Initializing IPFS daemon with the lowpower profile.\\n" "${INFO}"
+#                    sudo -u $USER_ACCOUNT ipfs init --profile=lowpower
+#                else
                     sudo -u $USER_ACCOUNT ipfs init
-                fi
-            fi
+#                fi
+#            fi
 
         fi
 
@@ -10424,22 +10424,22 @@ ipfs_update_port() {
         printf " =============== Starting: IPFS ========================================\\n\\n"
         # ==============================================================================
 
-        if [ "$VERBOSE_MODE" = true ]; then
-            printf "%b Verbose Mode: DGB_NETWORK_FINAL - $DGB_NETWORK_FINAL\\n" "${INFO}"
-            printf "%b Verbose Mode: IPFS_PORT_IP4 - $IPFS_PORT_IP4\\n" "${INFO}"
-            printf "%b Verbose Mode: IPFS_PORT_IP6 - $IPFS_PORT_IP6\\n" "${INFO}"
-            printf "%b Verbose Mode: IPFS_PORT_IP4_QUIC - $IPFS_PORT_IP4_QUIC\\n" "${INFO}"
-            printf "%b Verbose Mode: IPFS_PORT_IP6_QUIC - $IPFS_PORT_IP6_QUIC\\n" "${INFO}"
-            printf "\\n"
-        fi
-
-        printf "%b IPFS is installed but not currently running.\\n" "${INFO}"
+#        if [ "$VERBOSE_MODE" = true ]; then
+#            printf "%b Verbose Mode: DGB_NETWORK_FINAL - $DGB_NETWORK_FINAL\\n" "${INFO}"
+#            printf "%b Verbose Mode: IPFS_PORT_IP4 - $IPFS_PORT_IP4\\n" "${INFO}"
+#            printf "%b Verbose Mode: IPFS_PORT_IP6 - $IPFS_PORT_IP6\\n" "${INFO}"
+#            printf "%b Verbose Mode: IPFS_PORT_IP4_QUIC - $IPFS_PORT_IP4_QUIC\\n" "${INFO}"
+#            printf "%b Verbose Mode: IPFS_PORT_IP6_QUIC - $IPFS_PORT_IP6_QUIC\\n" "${INFO}"
+#            printf "\\n"
+#        fi
 
     if [ -f "$USER_HOME/.ipfs/config" ]; then
 
         # If using DigiByte testnet, change default Kubo IPFS port to 4004
 
         local update_ipfsport_now
+
+        printf "%b Checking Kubo IPFS ports...\\n" "${INFO}"
 
         if [[ "$DGB_NETWORK_FINAL" = "TESTNET" ]] && [[ "$IPFS_PORT_IP4" = "4001" ]]; then
             printf "%b Using DigiByte testnet. Updating Kobo IPFS ports...\\n" "${INFO}"
@@ -10532,6 +10532,8 @@ ipfs_update_port() {
     if [ -f "$USER_HOME/.jsipfs/config" ]; then
 
         # If using DigiByte testnet, change default JS-IPFS port to 4004
+
+        printf "%b Checking JS-IPFS ports...\\n" "${INFO}"
 
         local update_ipfsport_now
 
@@ -11213,7 +11215,6 @@ if [ "$NODEJS_DO_INSTALL" = "YES" ]; then
     if [ "$NODEJS_VER_LOCAL" = "" ]; then
         NODEJS_VER_LOCAL=$(node -v 2>/dev/null | sed 's/v//g')
     fi
-
 
     # Update diginode.settings with new Node.js local version number and the install/upgrade date
     sed -i -e "/^NODEJS_VER_LOCAL=/s|.*|NODEJS_VER_LOCAL=\"$NODEJS_VER_LOCAL\"|" $DGNT_SETTINGS_FILE